Rename External to LinkedInAPI
